+++ sensor firmware/get_mic_data_and_send_fft/sending_queue.cpp
@@ -14,11 +14,11 @@ int free_size = 0;
 
 
 
-void init_sending_queue(){
+void init_sending_queue() {
   sending_list *pool;
   pool = (sending_list*)heap_caps_malloc(sizeof(sending_list) * 64, MALLOC_CAP_8BIT);
 
-  for(int i = 0; i<64; i++){
+  for (int i = 0; i < 64; i++) {
     pool[i].next = free_pool;
     free_pool = &pool[i];
   }
@@ -26,8 +26,8 @@ void init_sending_queue(){
 
 
 
-void need_free_items(){
-  if(heap_caps_get_free_size(MALLOC_CAP_8BIT) < 50000){
+void need_free_items() {
+  if (heap_caps_get_free_size(MALLOC_CAP_8BIT) < 50000) {
     remove_first();
     Serial.println((unsigned long) first_item);
   } else {
@@ -36,15 +36,15 @@ void need_free_items(){
 }
 
 
-sending_list * get_one_free_item_and_add_to_queue(){
+sending_list * get_one_free_item_and_add_to_queue() {
   sending_list *new_item = free_pool;
   free_pool = (sending_list*)new_item->next;
   return new_item;
 }
 
 
-void put_item_into_queue(sending_list * new_item){
-  if (!first_item){
+void put_item_into_queue(sending_list * new_item) {
+  if (!first_item) {
     first_item = new_item;
     last_item = new_item;
   } else {
@@ -58,13 +58,13 @@ void put_item_into_queue(sending_list * new_item){
 
 
 
-void add_to_sending_queue(double* fft, double decibels, long timestamp){
-  if(!free_pool){
+void add_to_sending_queue(double* fft, double decibels, unsigned int timestamp) {
+  if (!free_pool) {
     need_free_items();
   }
   sending_list *new_item = get_one_free_item_and_add_to_queue();
 
-  for(int i = 0; i< DOWNSAMPLED__FFT; i++){
+  for (int i = 0; i < DOWNSAMPLED__FFT; i++) {
     new_item->fft_values[i] = fft[i];
   }
   new_item->next = (sending_list*)0;
@@ -79,14 +79,14 @@ void add_to_sending_queue(double* fft, double decibels, long timestamp){
 
 
 
-sending_list * get_first(){
+sending_list * get_first() {
   return (sending_list *)first_item;
 }
 
 
 
-void remove_first(){
-  if(first_item){
+void remove_first() {
+  if (first_item) {
     in_list --;
     sending_list* second = (sending_list*)first_item->next;
     first_item->next = free_pool;
