+++ sensor firmware/get_mic_data_and_send_fft/sending_queue.h
@@ -1,8 +1,8 @@
 #ifndef SENDING_QUEUE_H
 #define SENDING_QUEUE_H
-
+#include <arduino.h>
 #include "global_defines.h"
-void add_to_sending_queue(double* fft, double decibels, long timestamp);
+void add_to_sending_queue(double* fft, double decibels, unsigned int timestamp);
 sending_list * get_first();
 void remove_first();
